+++ homeassistant/custom_components/dsc_hub/frontend/src/components/index.ts
@@ -1,3 +1,3 @@
 export { SoilTestWizard } from "./SoilTestWizard";
 export type { SoilTestWizardProps } from "./SoilTestWizard";
-export { SankeyFlowPrototype } from "./SankeyFlowPrototype";
+export { FlowSankey, SankeyFlowPrototype } from "./FlowSankey";